Make our own loop implementation which is much simpler
We simply forget about the blocking shenanigans from psychopy.session (Hint: it
is highly broken). And we added our own async function to expose to FastAPI,
using future to control asynchronous operation between the main thread (that
will be waiting on operation from the server thread) and the serving thread.

It seems to work. at least once.